fixed most remaining warnings

+++ src/framework/image/tga.cpp
@@ -67,8 +67,7 @@ void TGAHeader::save(std::ostream& stream)
    }
 }
 
-
-int savetga(const std::string& filename, uint32_t* data, int32_t width, int32_t height)
+int savetga(const std::string& filename, uint32_t* data, int16_t width, int16_t height)
 {
    TGAHeader header(width, height, 32);
    std::ofstream stream(filename, std::ios::binary);
